Refactor all integration tests
