+++ integration/src/main/java/scripts/testscripts/ControlledBigQueryDatasetLifecycle.java
@@ -1,7 +1,10 @@
 package scripts.testscripts;
 
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.greaterThanOrEqualTo;
+import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -10,10 +13,16 @@ import bio.terra.testrunner.runner.config.TestUserSpecification;
 import bio.terra.workspace.api.ControlledGcpResourceApi;
 import bio.terra.workspace.api.WorkspaceApi;
 import bio.terra.workspace.client.ApiException;
+import bio.terra.workspace.model.CloneControlledGcpBigQueryDatasetRequest;
+import bio.terra.workspace.model.CloneControlledGcpBigQueryDatasetResult;
+import bio.terra.workspace.model.ClonedControlledGcpBigQueryDataset;
+import bio.terra.workspace.model.CloningInstructionsEnum;
 import bio.terra.workspace.model.GcpBigQueryDatasetResource;
 import bio.terra.workspace.model.GcpBigQueryDatasetUpdateParameters;
 import bio.terra.workspace.model.GrantRoleRequestBody;
 import bio.terra.workspace.model.IamRole;
+import bio.terra.workspace.model.JobControl;
+import bio.terra.workspace.model.ResourceMetadata;
 import bio.terra.workspace.model.UpdateControlledGcpBigQueryDatasetRequestBody;
 import com.google.api.client.http.HttpStatusCodes;
 import com.google.cloud.bigquery.BigQuery;
@@ -21,6 +30,8 @@ import com.google.cloud.bigquery.BigQueryException;
 import com.google.cloud.bigquery.Dataset;
 import com.google.cloud.bigquery.DatasetId;
 import com.google.cloud.bigquery.Field;
+import com.google.cloud.bigquery.FieldValue;
+import com.google.cloud.bigquery.FieldValueList;
 import com.google.cloud.bigquery.Job;
 import com.google.cloud.bigquery.JobId;
 import com.google.cloud.bigquery.JobInfo;
@@ -33,17 +44,18 @@ import com.google.cloud.bigquery.Table;
 import com.google.cloud.bigquery.TableId;
 import com.google.cloud.bigquery.TableInfo;
 import com.google.cloud.bigquery.TableResult;
+import java.time.Duration;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.StreamSupport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scripts.utils.BqDatasetUtils;
 import scripts.utils.ClientTestUtils;
-import scripts.utils.CloudContextMaker;
+import scripts.utils.GcpWorkspaceCloneTestScriptBase;
 import scripts.utils.SamClientUtils;
-import scripts.utils.WorkspaceAllocateTestScriptBase;
 
-public class ControlledBigQueryDatasetLifecycle extends WorkspaceAllocateTestScriptBase {
+public class ControlledBigQueryDatasetLifecycle extends GcpWorkspaceCloneTestScriptBase {
   private static final Logger logger = LoggerFactory.getLogger(ControlledGcsBucketLifecycle.class);
 
   private static final String DATASET_RESOURCE_NAME = "wsmtest_dataset";
@@ -77,7 +89,7 @@ public class ControlledBigQueryDatasetLifecycle extends WorkspaceAllocateTestScr
     ControlledGcpResourceApi ownerResourceApi =
         ClientTestUtils.getControlledGcpResourceClient(testUser, server);
 
-    // Add a writer and reader to this workspace
+    // Add a writer and reader to the source workspace
     logger.info(
         "Adding {} as writer and {} as reader to workspace {}",
         writer.userEmail,
@@ -87,12 +99,13 @@ public class ControlledBigQueryDatasetLifecycle extends WorkspaceAllocateTestScr
         new GrantRoleRequestBody().memberEmail(writer.userEmail), getWorkspaceId(), IamRole.WRITER);
     workspaceApi.grantRole(
         new GrantRoleRequestBody().memberEmail(reader.userEmail), getWorkspaceId(), IamRole.READER);
+    // Also add the "reader" user as a writer to the destination workspace for cloning.
+    workspaceApi.grantRole(
+        new GrantRoleRequestBody().memberEmail(reader.userEmail),
+        getDestinationWorkspaceId(),
+        IamRole.WRITER);
 
     SamClientUtils.dumpResourcePolicy(testUser, server, "workspace", getWorkspaceId().toString());
-
-    // Create a cloud context
-    String projectId = CloudContextMaker.createGcpCloudContext(getWorkspaceId(), workspaceApi);
-    logger.info("Created project {}", projectId);
 
     // Create a shared BigQuery dataset
     GcpBigQueryDatasetResource createdDataset =
@@ -114,12 +127,12 @@ public class ControlledBigQueryDatasetLifecycle extends WorkspaceAllocateTestScr
 
     createControlledDatasetWithBothResourceNameAndDatasetIdSpecified(ownerResourceApi);
 
-    BigQuery ownerBqClient = ClientTestUtils.getGcpBigQueryClient(testUser, projectId);
-    BigQuery writerBqClient = ClientTestUtils.getGcpBigQueryClient(writer, projectId);
-    BigQuery readerBqClient = ClientTestUtils.getGcpBigQueryClient(reader, projectId);
+    BigQuery ownerBqClient = ClientTestUtils.getGcpBigQueryClient(testUser, getSourceProjectId());
+    BigQuery writerBqClient = ClientTestUtils.getGcpBigQueryClient(writer, getSourceProjectId());
+    BigQuery readerBqClient = ClientTestUtils.getGcpBigQueryClient(reader, getSourceProjectId());
 
     // Workspace owner can create a table in this dataset
-    Table table = createTable(ownerBqClient, projectId);
+    Table table = createTable(ownerBqClient, getSourceProjectId());
     String tableName = table.getTableId().getTable();
 
     // Workspace reader can read the table
@@ -188,7 +201,8 @@ public class ControlledBigQueryDatasetLifecycle extends WorkspaceAllocateTestScr
     assertEquals(HttpStatusCodes.STATUS_CODE_BAD_REQUEST, invalidUpdateEx.getCode());
 
     // Cloud metadata matches the updated values
-    Dataset cloudDataset = ownerBqClient.getDataset(DatasetId.of(projectId, DATASET_RESOURCE_NAME));
+    Dataset cloudDataset =
+        ownerBqClient.getDataset(DatasetId.of(getSourceProjectId(), DATASET_RESOURCE_NAME));
     assertEquals(defaultTableLifetimeSec * 1000L, cloudDataset.getDefaultTableLifetime());
     assertNull(cloudDataset.getDefaultPartitionExpirationMs());
 
@@ -197,7 +211,15 @@ public class ControlledBigQueryDatasetLifecycle extends WorkspaceAllocateTestScr
         "Deleting table {} from dataset {}", table.getTableId().getTable(), DATASET_RESOURCE_NAME);
     assertTrue(
         writerBqClient.delete(
-            TableId.of(projectId, DATASET_RESOURCE_NAME, table.getTableId().getTable())));
+            TableId.of(
+                getSourceProjectId(), DATASET_RESOURCE_NAME, table.getTableId().getTable())));
+
+    // Populate dataset with additional tables to verify cloning behavior
+    BqDatasetUtils.populateBigQueryDataset(createdDataset, testUser, getSourceProjectId());
+    // Verify workspace reader is able to clone the resource they can read
+    ControlledGcpResourceApi readerResourceApi =
+        ClientTestUtils.getControlledGcpResourceClient(reader, server);
+    testCloneBigQueryDataset(createdDataset, reader, readerResourceApi);
 
     // Workspace writer cannot delete the dataset directly
     var writerCannotDeleteException =
@@ -290,5 +312,127 @@ public class ControlledBigQueryDatasetLifecycle extends WorkspaceAllocateTestScr
       throw new RuntimeException(queryJob.getStatus().getError().toString());
     }
     return completedJob.getQueryResults();
+  }
+
+  private void testCloneBigQueryDataset(
+      GcpBigQueryDatasetResource sourceDataset,
+      TestUserSpecification cloningUser,
+      ControlledGcpResourceApi cloningUserResourceApi)
+      throws Exception {
+    final String destinationDatasetName =
+        ("clone_" + UUID.randomUUID().toString()).replace('-', '_');
+    // clone the dataset as the cloning user
+    final String clonedDatasetDescription = "Clone of " + destinationDatasetName;
+    final String jobId = UUID.randomUUID().toString();
+    final CloneControlledGcpBigQueryDatasetRequest cloneRequest =
+        new CloneControlledGcpBigQueryDatasetRequest()
+            .cloningInstructions(CloningInstructionsEnum.RESOURCE)
+            .description(clonedDatasetDescription)
+            .location(null) // keep same
+            .destinationWorkspaceId(getDestinationWorkspaceId())
+            .name("MyClonedDataset")
+            .jobControl(new JobControl().id(jobId))
+            .destinationDatasetName(null); // keep same
+    final ResourceMetadata sourceDatasetMetadata = sourceDataset.getMetadata();
+    logger.info(
+        "Cloning BigQuery dataset\n\tname: {}\n\tresource ID: {}\n\tworkspace: {}\n\t"
+            + "projectID: {}\ninto destination \n\tname: {}\n\tworkspace: {}\n\tprojectID: {}",
+        sourceDatasetMetadata.getName(),
+        sourceDatasetMetadata.getResourceId(),
+        sourceDatasetMetadata.getWorkspaceId(),
+        getSourceProjectId(),
+        sourceDatasetMetadata.getName(),
+        getDestinationWorkspaceId(),
+        getDestinationProjectId());
+
+    // Submit clone request and poll for async result
+    CloneControlledGcpBigQueryDatasetResult cloneResult =
+        cloningUserResourceApi.cloneBigQueryDataset(
+            cloneRequest,
+            sourceDatasetMetadata.getWorkspaceId(),
+            sourceDatasetMetadata.getResourceId());
+    cloneResult =
+        ClientTestUtils.pollWhileRunning(
+            cloneResult,
+            () ->
+                cloningUserResourceApi.getCloneBigQueryDatasetResult(
+                    cloneRequest.getDestinationWorkspaceId(), cloneRequest.getJobControl().getId()),
+            CloneControlledGcpBigQueryDatasetResult::getJobReport,
+            Duration.ofSeconds(5));
+
+    ClientTestUtils.assertJobSuccess(
+        "clone BigQuery dataset", cloneResult.getJobReport(), cloneResult.getErrorReport());
+    assertEquals(
+        sourceDatasetMetadata.getWorkspaceId(), cloneResult.getDataset().getSourceWorkspaceId());
+    assertEquals(
+        sourceDatasetMetadata.getResourceId(), cloneResult.getDataset().getSourceResourceId());
+
+    // unwrap the result one layer at a time
+    final ClonedControlledGcpBigQueryDataset clonedControlledGcpBigQueryDataset =
+        cloneResult.getDataset();
+    assertEquals(
+        CloningInstructionsEnum.RESOURCE,
+        clonedControlledGcpBigQueryDataset.getEffectiveCloningInstructions());
+
+    final GcpBigQueryDatasetResource clonedResource =
+        clonedControlledGcpBigQueryDataset.getDataset();
+    final ResourceMetadata clonedDatasetMetadata = clonedResource.getMetadata();
+    assertEquals(
+        sourceDatasetMetadata.getCloningInstructions(),
+        clonedDatasetMetadata.getCloningInstructions());
+    assertEquals(
+        sourceDatasetMetadata.getCloudPlatform(), clonedDatasetMetadata.getCloudPlatform());
+    assertEquals(sourceDatasetMetadata.getResourceType(), clonedDatasetMetadata.getResourceType());
+    assertEquals(
+        sourceDatasetMetadata.getStewardshipType(), clonedDatasetMetadata.getStewardshipType());
+    assertEquals(
+        sourceDatasetMetadata.getControlledResourceMetadata().getManagedBy(),
+        clonedDatasetMetadata.getControlledResourceMetadata().getManagedBy());
+    assertEquals(
+        sourceDatasetMetadata.getControlledResourceMetadata().getAccessScope(),
+        clonedDatasetMetadata.getControlledResourceMetadata().getAccessScope());
+    assertNotEquals(
+        sourceDataset.getAttributes().getProjectId(),
+        clonedResource.getAttributes().getProjectId());
+    assertEquals(
+        sourceDataset.getAttributes().getDatasetId(),
+        clonedResource.getAttributes().getDatasetId());
+    assertEquals(
+        sourceDataset.getMetadata().getName(), sourceDataset.getAttributes().getDatasetId());
+
+    // compare dataset contents
+    final BigQuery bigQueryClient =
+        ClientTestUtils.getGcpBigQueryClient(cloningUser, getDestinationProjectId());
+    final QueryJobConfiguration employeeQueryJobConfiguration =
+        QueryJobConfiguration.newBuilder(
+                "SELECT * FROM `"
+                    + getDestinationProjectId()
+                    + "."
+                    + clonedResource.getAttributes().getDatasetId()
+                    + ".employee`;")
+            .build();
+    final TableResult employeeTableResult = bigQueryClient.query(employeeQueryJobConfiguration);
+    final long numRows =
+        StreamSupport.stream(employeeTableResult.getValues().spliterator(), false).count();
+    assertThat(numRows, is(greaterThanOrEqualTo(2L)));
+
+    final TableResult departmentTableResult =
+        bigQueryClient.query(
+            QueryJobConfiguration.newBuilder(
+                    "SELECT * FROM `"
+                        + getDestinationProjectId()
+                        + "."
+                        + clonedResource.getAttributes().getDatasetId()
+                        + ".department` "
+                        + "WHERE department_id = 201;")
+                .build());
+    final FieldValueList row =
+        StreamSupport.stream(departmentTableResult.getValues().spliterator(), false)
+            .findFirst()
+            .orElseThrow(() -> new RuntimeException("Can't find expected result row"));
+    final FieldValue nameFieldValue = row.get("name");
+    assertEquals("ocean", nameFieldValue.getStringValue());
+    final FieldValue managerFieldValue = row.get("manager_id");
+    assertEquals(101, managerFieldValue.getLongValue());
   }
 }
