--- integration/src/main/resources/configs/integration/CloneBigQueryDataset.json
@@ -1,20 +1,0 @@
-{
-  "name": "CloneBigQueryDataset",
-  "description": "Clone test",
-  "serverSpecificationFile": "workspace-local.json",
-  "kubernetes": {},
-  "application": {},
-  "testScripts": [
-    {
-      "name": "CloneBigQueryDataset",
-      "numberOfUserJourneyThreadsToRun": 1,
-      "userJourneyThreadPoolSize": 1,
-      "expectedTimeForEach": 15,
-      "expectedTimeForEachUnit": "MINUTES",
-      "parameters": {
-        "spend-profile-id": "wm-default-spend-profile"
-      }
-    }
-  ],
-  "testUserFiles": ["bella.json", "elijah.json"]
-}
