--- integration/src/main/resources/configs/integration/CloneReferencedResources.json
@@ -1,27 +1,0 @@
-{
-  "name": "CloneReferencedResources",
-  "description": "Exercise reference access validation over all resource types",
-  "serverSpecificationFile": "workspace-local.json",
-  "kubernetes": {},
-  "application": {},
-  "testScripts": [
-    {
-      "name": "CloneReferencedResources",
-      "numberOfUserJourneyThreadsToRun": 1,
-      "userJourneyThreadPoolSize": 1,
-      "expectedTimeForEach": 5,
-      "expectedTimeForEachUnit": "MINUTES",
-      "parameters": {
-        "spend-profile-id": "wm-default-spend-profile",
-        "data-repo-instance": "terra",
-        "snapshot-id": "97b5559a-2f8f-4df3-89ae-5a249173ee0c",
-        "gcs-uniform-bucket": "gs://terra_wsm_test_resource",
-        "gcs-object": "gs://terra_wsm_fine_grained_test_bucket/foo/monkey_sees_monkey_dos.txt",
-        "gcs-folder": "gs://terra_wsm_fine_grained_test_bucket/foo/",
-        "bq-table": "projects/terra-kernel-k8s/datasets/terra_wsm_test_dataset/tables/terra wsm test data table",
-        "git-repo": "git@github.com:DataBiosphere/terra-workspace-manager.git"
-      }
-    }
-  ],
-  "testUserFiles": ["bella.json"]
-}
