+++ integration/src/main/resources/configs/integration/ReferencedGcsResourceLifecycle.json
@@ -1,25 +1,22 @@
 {
-  "name": "ValidateReferencedResources",
-  "description": "Exercise reference access validation over all resource types",
+  "name": "ReferencedGcsResourceLifecycle",
+  "description": "CRUD GCS bucket and file operations",
   "serverSpecificationFile": "workspace-local.json",
   "kubernetes": {},
   "application": {},
   "testScripts": [
     {
-      "name": "ValidateReferencedResources",
+      "name": "ReferencedGcsResourceLifecycle",
       "numberOfUserJourneyThreadsToRun": 1,
       "userJourneyThreadPoolSize": 1,
       "expectedTimeForEach": 5,
       "expectedTimeForEachUnit": "MINUTES",
       "parameters": {
         "spend-profile-id": "wm-default-spend-profile",
-        "data-repo-instance": "terra",
-        "snapshot-id": "97b5559a-2f8f-4df3-89ae-5a249173ee0c",
         "gcs-bucket": "gs://terra_wsm_fine_grained_test_bucket",
         "gcs-uniform-bucket": "gs://terra_wsm_test_resource",
         "gcs-object": "gs://terra_wsm_fine_grained_test_bucket/foo/monkey_sees_monkey_dos.txt",
-        "gcs-folder": "gs://terra_wsm_fine_grained_test_bucket/foo/",
-        "bq-table": "projects/terra-kernel-k8s/datasets/terra_wsm_test_dataset/tables/terra wsm test data table"
+        "gcs-folder": "gs://terra_wsm_fine_grained_test_bucket/foo/"
       }
     }
   ],
