--- integration/src/main/resources/configs/integration/UpdateReferenceResources.json
@@ -1,29 +1,0 @@
-{
-  "name": "UpdateReferenceResources",
-  "description": "Update resource reference for all the resource type",
-  "serverSpecificationFile": "workspace-local.json",
-  "kubernetes": {},
-  "application": {},
-  "testScripts": [
-    {
-      "name": "UpdateReferenceResources",
-      "numberOfUserJourneyThreadsToRun": 1,
-      "userJourneyThreadPoolSize": 1,
-      "expectedTimeForEach": 5,
-      "expectedTimeForEachUnit": "MINUTES",
-      "parameters": {
-        "spend-profile-id": "wm-default-spend-profile",
-        "data-repo-instance": "terra",
-        "snapshot-id": "97b5559a-2f8f-4df3-89ae-5a249173ee0c",
-        "snapshot-id-2": "de189c9b-ef96-4ade-9970-5829056083e3",
-        "gcs-uniform-bucket": "gs://terra_wsm_test_resource",
-        "gcs-object": "gs://terra_wsm_fine_grained_test_bucket/foo/monkey_sees_monkey_dos.txt",
-        "gcs-folder": "gs://terra_wsm_fine_grained_test_bucket/foo/",
-        "bq-table": "projects/terra-kernel-k8s/datasets/terra_wsm_test_dataset/tables/terra wsm test data table",
-        "bq-table-2": "projects/terra-kernel-k8s/datasets/terra_wsm_test_dataset_2/tables/terra wsm test data table 2",
-        "git-repo": "git@github.com:DataBiosphere/terra-workspace-manager.git"
-      }
-    }
-  ],
-  "testUserFiles": ["bella.json", "elijah.json"]
-}
