+++ integration/src/main/resources/testusers/README.md
@@ -28,5 +28,5 @@ Currently, users have the following access permissions, replicated across enviro
 |                                           | TDR Snapshot       | `wm-default-spend-profile` | BQ Dataset `terra_wsm_test_dataset` | BQ Dataset `terra_wsm_test_dataset_2` | BQ Data table `terra_wsm_test_dataset/terra wsm test data table` | BQ Data table `terra_wsm_test_dataset/terra wsm test data table 2` | BQ Data table `terra_wsm_test_dataset_2/terra wsm test data table 2` | Bucket `terra_wsm_test_resource` | Bucket `terra_wsm_fine_grained_test_bucket` | Bucket object `terra_wsm_fine_grained_test_bucket/foo/` | Bucket object `terra_wsm_fine_grained_test_bucket/foo/monkey_sees_monkey_dos.txt` |
 | ----------------------------------------- | ------------------ | -------------------------- | ----------------------------------- | ------------------------------------  | ---------------------------------------------------------------- | ------------------------------------------------------------------ | -------------------------------------------------------------------- | -------------------------------- | ------------------------------------------- | ------------------------------------------------------- | --------------------------------------------------------------------------------- |
 | **bella.redwalker**                       | :white_check_mark: | :white_check_mark:         | :white_check_mark:                  | :white_check_mark:                    | :white_check_mark:                                               | :white_check_mark:                                                 | :white_check_mark:                                                   | :white_check_mark:               | :white_check_mark:                          | :white_check_mark:                                      | :white_check_mark:                                                                |
-| **elijah.thunderlord**                    | :x:                | :white_check_mark:         | :x:                                 | :white_check_mark:                    | :x:                                                              | :x:                                                                | :white_check_mark:                                                                  | :white_check_mark:               | :white_check_mark:                          | :x:                                                     | :white_check_mark:                                                                |
+| **elijah.thunderlord**                    | :x:                | :white_check_mark:         | :x:                                 | :white_check_mark:                    | :x:                                                              | :x:                                                                | :white_check_mark:                                                   | :white_check_mark:               | :x:                                         | :x:                                                     | :white_check_mark:                                                                |
 | **liam.dragonmaw**                        | :x:                | :x:                        | :x:                                 | :x:                                   | :x:                                                              | :x:                                                                | :x:                                                                  | :x:                              | :x:                                         | :x:                                                     | :x:                                                                               |
